+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.3] - 2026-07-02
+
 ### Fixed
 
 - **pfSense (Plus) boxes stored with the wrong device type self-heal on agent